Spark: Move connector classes to avoid conflicts
