+++ spark3-extensions/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/Call.scala
@@ -21,7 +21,7 @@ package org.apache.spark.sql.catalyst.plans.logical
 
 import org.apache.spark.sql.catalyst.expressions.{Attribute, Expression}
 import org.apache.spark.sql.catalyst.util.truncatedString
-import org.apache.spark.sql.connector.catalog.Procedure
+import org.apache.spark.sql.connector.iceberg.catalog.Procedure
 import scala.collection.Seq
 
 case class Call(procedure: Procedure, args: Seq[Expression]) extends Command {
