+++ spark3/src/main/java/org/apache/spark/sql/connector/iceberg/catalog/ProcedureParameter.java
@@ -17,19 +17,21 @@
  * under the License.
  */
 
-package org.apache.spark.sql.connector.catalog;
+package org.apache.spark.sql.connector.iceberg.catalog;
 
-import org.apache.spark.sql.catalyst.InternalRow;
-import org.apache.spark.sql.types.StructType;
+import org.apache.spark.sql.types.DataType;
 
-public interface Procedure {
-  ProcedureParameter[] parameters();
+public interface ProcedureParameter {
 
-  StructType outputType();
-
-  InternalRow[] call(InternalRow input);
-
-  default String description() {
-    return this.getClass().toString();
+  static ProcedureParameter required(String name, DataType dataType) {
+    return new ProcedureParameterImpl(name, dataType, true);
   }
+
+  static ProcedureParameter optional(String name, DataType dataType) {
+    return new ProcedureParameterImpl(name, dataType, false);
+  }
+
+  String name();
+  DataType dataType();
+  boolean required();
 }
